Updated __init__.py so that it will automatically update the added modules and deleted modules.py
