+++ pygorithm/__init__.py
@@ -30,7 +30,7 @@ Timothy 'Tjsretchalot' Moore
 
 """
 
-__version__ = '1.0.0'
+__version__ = '1.0.3'
 __author__ = 'Omkar Pathak'
 
 # List maintainers here
@@ -64,4 +64,5 @@ __all__ = [
     'sorting',
     'string',
     'pathfinding',
+    'geometry',
 ]
