+++ pygorithm/data_structures/__init__.py
@@ -1,19 +1,21 @@
 """
 Collection of data structure examples
 """
-from . import graph
-from . import heap
-from . import linked_list
-from . import queue
-from . import stack
-from . import tree
+import pkgutil
 
-__all__ = [
-    'graph',
-    'heap',
-    'linked_list',
-    'queue',
-    'stack',
-    'tree',
-    'trie'
-]
+def modules():
+    """
+    Find all functions in pygorithm.data_structures
+    """
+    from pygorithm import data_structures
+    package = data_structures
+    modules_list = []
+    for importer, modname, ispkg in pkgutil.iter_modules(package.__path__):
+        modules_list.append(modname)
+    modules_list.remove('modules')
+    modules_list.sort()
+    return modules_list
+
+modules_list = modules()
+
+__all__ = modules_list
