+++ pygorithm/fibonacci/__init__.py
@@ -1,15 +1,21 @@
 """
 Collection of fibonacci methods and functions
 """
-from . import generator
-from . import goldenratio
-from . import memoization
-from . import modules
-from . import recursion
+import pkgutil
 
-__all__ = [
-    'generator',
-    'goldenratio',
-    'memoization',
-    'recursion'
-]
+def modules():
+    """
+    Find all functions in pygorithm.fibonacci
+    """
+    from pygorithm import fibonacci
+    package = fibonacci
+    modules_list = []
+    for importer, modname, ispkg in pkgutil.iter_modules(package.__path__):
+        modules_list.append(modname)
+    modules_list.remove('modules')
+    modules_list.sort()
+    return modules_list
+
+modules_list = modules()
+
+__all__ = modules_list
