+++ pygorithm/geometry/__init__.py
@@ -1,8 +1,21 @@
 """
 Collection of special geometry functions
 """
-from . import rect_broad_phase
+import pkgutil
 
-__all__ = [
-    'collision_detection.py'
-]
+def modules():
+    """
+    Find all functions in pygorithm.geometry
+    """
+    from pygorithm import geometry
+    package = geometry
+    modules_list = []
+    for importer, modname, ispkg in pkgutil.iter_modules(package.__path__):
+        modules_list.append(modname)
+    modules_list.remove('modules')
+    modules_list.sort()
+    return modules_list
+
+modules_list = modules()
+
+__all__ = modules_list
