+++ pygorithm/math/__init__.py
@@ -1,12 +1,21 @@
 """
 Collection of special math functions
 """
-from . import lcm
-from . import lcm_using_gcd
-from . import sieve_of_eratosthenes
+import pkgutil
 
-__all__ = [
-    'lcm',
-    'lcm_using_gcd',
-    'sieve_of_eratosthenes'
-]
+def modules():
+    """
+    Find all functions in pygorithm.math
+    """
+    from pygorithm import math
+    package = math
+    modules_list = []
+    for importer, modname, ispkg in pkgutil.iter_modules(package.__path__):
+        modules_list.append(modname)
+    modules_list.remove('modules')
+    modules_list.sort()
+    return modules_list
+
+modules_list = modules()
+
+__all__ = modules_list
